feat(olink): reduce network load for prop changes
Previously, every call to SetProperty immediately triggered a call to
the olink protocol and sent the change request to the service. This way
it was easily possible to flood the service with out providing any new
information by just keep sending already set values.
This change checks whether the value is already set on the service side
or whether it was already requested - and only if it is really a new
value we send the request to the service.
This adds a bit more performance load on the client side but keeps the
service responsive.