feat(jni): add starting android service
Introduce a service starter class that allows starting and stopping the
android service provided by the java template.

An android service needs to be explicitly started and in a thread of the
unreal application, JNI backend with unreal implementation must be
provided to it.  The java part of service starter of the JNI module,
handles most of that, the added cpp class runs that code.

The object returned by it is the java JNI service backend that was
instantiated for the service. It should be used to perform operations
on it.