feat(jni): service handles methods
Translates data from java to cpp and result from cpp to java.
Does not use template for the result to keep the name more meaningfull.
WARNING: the executing of method may cause some side effects like changing
the property or emitting a signal. In this case the application will
crash.
The way to solve this is either always broadcast with AsyncTask
or have a promise for the result, schedule executing method with
AsyncTask which should fill the result.
The method should wait for it.