feat(jni): client implement native is ready
When the jni client notifies it is ready to use (or changes that status
to false) it calls native function. This notification should be
broadcasted to whoever is interested about the connection status
changed.
For this purpose the delegates are added and exposed.
The commit adds those delegates along with a dummy class which is
necessary for the delegates to properly compile.
The commit introduces a mechanism of global function, that is provided by the
unreal client impl on init and set to an empty default version on
deinit.
Thanks to that function a more complexed actiom may be executed like
allow changing a member value (to track the state).
Also AsyncTask is used as the broadcasting informs also subscribers that
are expected to be in blueprints.