feat(jni): client handles native properties change
Introduces a global function for each property that gets changed.
The function by default just logs, but is set but the unreal client on
init and reset to the empty one on deinit.

The behavior added by unreal client is: to change the local value and
broadcast the value. The fact that it needs change the object state
makes it impossible to just use the public api of the object.