feat(jni): client handles signals
It uses the global pointer to the unreal client class, which is set in
the init and reset in deinit.
Since the signals are only to be broadcasted and no internal state
needs to be changed it was enough just to expose the unreal client, no
need for extra functions.