feat(jni): client helpers for execute methods
Executing a methods requires preparing a promise and storing it while
waiting for the message reaches the service and a response is delivered
back to the client.

This commit provides a method helper which allow storing promises for
certain result types.

This object is stored globally (the client is anyway a subsystem).