feat(jni): client helpers for execute methods
Executing a methods requires preparig a promise and storying it while
waiting for the messege reaches the service and a response is delivered
back to the client.
This commit provides a method helper which allow storing promisses for
certain result types.
This object is stored globally (the client is anyway a subsystem).