feat(jni): extend the config to use java code
Fill in the upl file to copy necessary java modules. The modules must be
placed inside the plugin directory in the android folder. This code is
generated by java template.

The also expose the services provided by the java code, adds permission
for binding, adds queries (this part is necessary for clients).

All the dependencies for includes and modules are also added.