feat(jni): extend the config to use java code
fill in the upl file to copy necessary java modules.
the modules must be placed inside the plugin directory in the android
folder. This code is generated by java template.
The also expose the services provided by the java code, adds permission for binding, adds queries (this part is necessery for clients).
All the dependencies for includes and modules are also added.