refactor(jni): move java generated code
It was expected for the jni feature, that java generated code will be in
android folder in each module.

This caused complex setup for unreal solution with more than one module,
with this feature all generated java code should be put next to the
folder in which the plugins are generated.