feat(jni): add class to cache jni ids
Add a central cache to store all the jclasses, jFields and jMethods -
acquire them on initialize.
They are known at that time and they don't change through the lifetime
of program.
In next commits there will be added checking for jni exception.
This cache is later used in adapters and data converters, the necessery
java types are filled, and will be checked for errors.