chore(jni): remove invlid comment
The motivation for it is template facitlitation. The string (or string arrays) are the only java types that are handled with wrapeer, using those directly requires another ifs for using the variable, condition to check if local ref needs to be removed. Also it would require adding new handling for getters, which are not allowed to delete the local/scoped object as it needs to be passed to java.