fix(jni): handle jclass resources in converter
jclass references are global references that needs to be deleted when no longer used.
It is also costly operation to get those, and they best be stored as statics, searched once, and reused.
With our current build model, the java libs are not dynamic, they don't change, so once loaded it is safe to use them through the lifetime of the program.
There is an issue when to initialize them. The startup of  the module is too early (it is cleaned up at the shutdown though).
It may be introduced in the module without any interface (structs and enums only module), also it must be ready when any other module want's to use it.
Hence I introduced the one-time thread-safe lazy initialization for those JNI resources. It must be checked on each function use.