fix(jni): clean up string local ref
In a ticket APIGEAR-372 there was suspicion of different handling
conversion of strings with unreal jni helper.
This tuned out to be not true (although 4.27 does not convert string
arrays - there is different ticket to align to 4.27 APIGEAR-401).
All versions convert the strings but putting it in wrappers that manage
local ref lifetime. For template facitlitation we in most cases we make the localref anyway. Using the wrapped ones directly requires another ifs for using the variable, condition to check if local ref needs to be removed. Also it requires adding new handling for getters, which are not allowed to delete the local/scoped object as it needs to be passed to java. Hence the decision to make local ref for simpler handling. This may be changed.
Anyway in data converter when handling the field of string type, there
was local ref made, but never deleted. This was made in scope only for
string handling, so it can easy be used with the wrapper only, no need
for local ref, no need for removing it later then.