fix(jni): make adapter java jni calls thread safe
Native calls from jni are always called from jni thread.
The Unreal's rule is that UObjects (including subsystems) should not be
used from threads other than game thread for most usages, due to
multithreading issues. But within current usage, and assuming:
- Adapter is an EngineSubsystem which lifetime is guaranteed for application lifetime
- when jni code is used, the editor is not used (jni code only runs on
  android devices, where editing blueprints is not possible)
- The backend is UPROPERTY and is thread-safe subsystem itself.
It is safety criteria are met.
The calls need also an easy access to the subsystem.
Although it is the jni adapter susbsystem's responsibility to start and
stop the java backend which implements those native calls, since the jni
has its separate thread some race conditions cannot be excluded,
especially during tear down.
The subsystems itself won't be deleted or garbage collected, but the
calls are not supposed to be used after deinitialize.
To provide the easy and access and improve the thread safety, espacially
on teardown the std::atomic was added around the handle, with which jni
native calls are using the subsystem.
To give the native calls access to only thing they need I introduced a
simple interface that is implemented by the subsystem.
The intention was to inherit it privately, but it seems unreal Subsystem
does not accept such form.
The handle is assinged in initialize and reset in deinitialize, which
prevents from using it outside the proper time slot.
In adapter the native jni calls operate on currently set backend.
The backend may change, but it is not expected to be frequent if at all,
nevertheless it should be guarded with a mutex, so the usage won't
occur while jni is asking for it.
The already existing UFUNCTION to get the backend is not guarded with
it, the assumption is that only game thread will change it, and that
this function is used from game thred - this means that guarding it with
mutex is not necessary.
WARNING: how about calling methods on backend in case the methods are
implemented in blueprints?